Add freebsd13 (AKA CURRENT) support
Add freebsd13 as supported platform; currently, libc works on freebsd13,
even if not supported by libc thanks to the implicit backward compatibility
provided by FreeBSD.

However, test fails and it stops me to add missing API to the crate.
This patch will solve the issues, properly defining freebsd13 as a new
target.
Unfortunately, cirrus-ci doesn't provide any VM based on the FreeBSD
development snapshot (13), so it's impossible atm to test this platform
with the automated pipeline